% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{databases}
\alias{databases}
\title{Get a database as a data frame}
\usage{
databases(id)
}
\arguments{
\item{id}{The ID of an individual database}
}
\description{
Returns all of the properties in a database as a data frame.
}
\examples{
\dontrun{
# Find the ID of a specific database
db_id <- search_all("Customer DB")$id

# Get that database
databases(db_id)
}
}
