% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_mrn_record_id}
\alias{get_mrn_record_id}
\title{get_mrn_record_id
. This function is used to check whether a given MRN is in REDcap.  It would be nice to do this
all in one go, but we can't both filter by a non-record-id field (in this case, the MRN)
and also return fields from instruments that lack the field we are filtering by.
In this case, the MRN is in the baseline pt info instrument, but the diet data is in the computrition instrument.
So, we have to first use this function to fetch the right record ID, then query with that record id
to get the actual meal data}
\usage{
get_mrn_record_id(mrn)
}
\arguments{
\item{mrn}{}
}
\value{
redcapr response object with fields in data for eb_mrn and record_id
}
\description{
get_mrn_record_id
. This function is used to check whether a given MRN is in REDcap.  It would be nice to do this
all in one go, but we can't both filter by a non-record-id field (in this case, the MRN)
and also return fields from instruments that lack the field we are filtering by.
In this case, the MRN is in the baseline pt info instrument, but the diet data is in the computrition instrument.
So, we have to first use this function to fetch the right record ID, then query with that record id
to get the actual meal data
}
\examples{
\dontrun{
get_mrn_record_id(12345)
}
}
