% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolkit.R
\name{merge_meals_and_units}
\alias{merge_meals_and_units}
\title{get patient data}
\usage{
merge_meals_and_units(dev_data, unittable, fndds_summary)
}
\arguments{
\item{dev_data}{}

\item{unittable}{}

\item{fndds_summary}{}
}
\value{
list
}
\description{
get patient data
}
