% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_all.R
\name{search_all}
\alias{search_all}
\title{Search for pages and databases}
\usage{
search_all(query = "")
}
\arguments{
\item{query}{A string to search pages and databases. An empty string returns
all pages and databases you have access to (this is the default).}
}
\description{
Specify a query to search all of the Notion pages and databases that you have
access to.
}
\examples{
\dontrun{
# Returns all pages and databases
search_all()

# Returns only pages and databases matching the query
search_all("Customer DB")
}
}
