% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolkit.R
\name{tabulate_pt_nutrition}
\alias{tabulate_pt_nutrition}
\title{tabulate patient nutrition}
\usage{
tabulate_pt_nutrition(pt_data_full_merge, mrn, dt_start, dt_end, nutrient_list)
}
\arguments{
\item{pt_data_full_merge}{output from merge_meals_and_units}
}
\description{
tabulate patient nutrition
}
