% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOSH-package.R
\docType{package}
\name{NOSH-package}
\alias{NOSH}
\alias{NOSH-package}
\title{NOSH: Shiny app for nutritional optimization for science and health}
\description{
Simplify interactions between nutrition management, clinical data, and FNDDS. NOSH facilitates diet data entry, curation, and redcap transactions.
}
\author{
\strong{Maintainer}: Nick Waters \email{nickp60@gmail.com}

Authors:
\itemize{
  \item John Slingerland \email{slingerj@mskcc.org}
  \item Natalie Smith \email{smithn7@mskcc.org}
  \item Anqi Dai \email{daia1@mskcc.org}
  \item Mirae Baichoo \email{daia1@mskcc.org}
}

}
\keyword{internal}
